integrated-backend/prioritization-ml/requirements.txt
# PyTorch Priority Classifier - Python Dependencies
# Install with: pip install -r requirements.txt

# Core ML
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Message Queue
pika>=1.3.0

# Utilities
joblib>=1.3.0
PyPDF2>=3.0.0
reportlab>=4.0.0
Pillow>=10.0.0

# Monitoring
prometheus_client>=0.17.0
